// src/types/StudentTypes/Values.ts
export interface Values {
  email: string;
  tel: string;
  firstName: string;
  lastName: string;
  githubUserName: string;
  bio: string;
  targetWorkCity: string;
  expectedSalary: number;
  monthsOfCommercialExp: number;
  education: string;
  workExperience: string;
  courses: string;
  portfolioInput: string;
  projectScrum: string;
  projectInput: string;
}